change method of choose page

+++ case.py
@@ -11,6 +11,9 @@ class Case(Driver):
         self.page = page
         case_path = self.load_json('./config/case_path.json')['path']
         self.path = os.path.join(case_path, f'{page.replace("/", "")}{name}.json')
+        if os.path.exists(self.path):
+            print('文件名已存在,为防覆盖请加name参数')
+            exit()
         self.fp = open(self.path, 'w', encoding='utf8')
 
         super(Case, self).__init__(url)
@@ -19,11 +22,12 @@ class Case(Driver):
         """
         写入json,关闭文件句柄和浏览器
         """
-        self.fp.write(json.dumps(self.res, ensure_ascii=False, indent=2))
-        self.fp.close()
-        if not self.res:
-            os.remove(self.path)
-        self.driver.close()
+        if getattr(self, "fp", False):
+            self.fp.write(json.dumps(self.res, ensure_ascii=False, indent=2))
+            self.fp.close()
+            if not self.res:
+                os.remove(self.path)
+            self.driver.close()
 
     def add_login(self, name='admin'):
         self.res.append({'login': name})
@@ -37,8 +41,8 @@ class Case(Driver):
     def add_close(self):
         self.res.append('close')
 
-    def add_choose(self, index=1):
-        self.res.append({'choose': index})
+    def add_choose(self, model, model_id):
+        self.res.append({'choose': {model: model_id}})
 
     def add_create(self):
         self.res.append('create')
@@ -75,17 +79,15 @@ class Case(Driver):
                 fill[label.text] = ''
         return fill
 
-    def add_login_wkf_logout(self, username, button, index=1):
+    def add_login_wkf_logout(self, username, button):
         """
         wkf流程:登录-打开-选择-通过/拒绝等-退出
         :param username: userinfo.json中的用户名
         :param button: wkf_button.json中的按键名
-        :param index: 索引
         """
         self.res.append({
             "login": username,
-            "open": self.page,
-            "choose": index,
+            "open": -1,
             "wkf": button,
             "logout": None
         })
@@ -106,17 +108,20 @@ class Case(Driver):
 
 
 if __name__ == '__main__':
-    c = Case('http://127.0.0.1:8069/web', page='还款/费用报销单(项目部及采购)', name='')
+    c = Case('http://127.0.0.1:8069/web', page='外协分包合同扣款登记表', name='factory')
     # c.add_login('test')
     # c.add_open()
     # c.add_create()
     # c.add_fill()
     # c.add_save()
-    c.add_login_create_logout(username="test")
+    c.add_login_create_logout(username="成本管理员")
+    # c.add_login_wkf_logout('生产', '通过')
+
     c.add_login_wkf_logout(username='采购', button='通过')
-    c.add_login_wkf_logout('财务', '通过')
-    c.add_login_wkf_logout('工厂', '通过')
-    c.add_login_wkf_logout('控股', '通过')
-    c.add_login_wkf_logout('test', '通过')
+    # c.add_login_wkf_logout('财务', '通过')
+    # c.add_login_wkf_logout('工厂', '通过')
+    # c.add_login_wkf_logout('控股', '通过')
+    c.add_login_wkf_logout('成本管理员', '通过')
+    # c.add_login_wkf_logout('test', '通过')
     # # c.add_login_wkf_logout('财务', '通过')
-    c.add_login_wkf_logout('工厂出纳', '通过')
+    # c.add_login_wkf_logout('工厂出纳', '通过')
